This addresses more obsolescence from issue #338
Obsolete methods that have been removed:

 - IKernel/DefaultKernel AddFacility(String key, IFacility facility)
 - IWindsorContainer/WindsorContainer AddFacility(String key, IFacility facility)
 - Facility configurations with an id attribute will yield a ConfigurationProcessingException
