+++ src/Castle.Facilities.WcfIntegration.Tests/Client/WcfClientFixture.cs
@@ -2133,8 +2133,7 @@ namespace Castle.Facilities.WcfIntegration.Tests
 		private readonly static string xmlConfiguration = @"<?xml version='1.0' encoding='utf-8' ?>
         <configuration>
         	<facilities>
-        		<facility id='wcf' 
-        				  type='Castle.Facilities.WcfIntegration.WcfFacility,
+        		<facility type='Castle.Facilities.WcfIntegration.WcfFacility,
         				        Castle.Facilities.WcfIntegration' />
         	</facilities>
